requirements.txt
discord
python-dotenv
paddlepaddle
setuptools
paddleocr
Pillow
gspread
oauth2client